Refaktorer ut gjenbrukbar komponent
