+++ app/src/features/InformasjonsseksjonMedKilde.tsx
@@ -1,0 +1,32 @@
+import { Detail, Heading } from "@navikt/ds-react";
+import clsx from "clsx";
+
+type InformasjonsseksjonMedKildeProps = {
+  kilde: string;
+  tittel: string;
+  children: React.ReactNode;
+  className?: string;
+};
+export const InformasjonsseksjonMedKilde = ({
+  children,
+  kilde,
+  tittel,
+  className,
+}: InformasjonsseksjonMedKildeProps) => {
+  return (
+    <div
+      className={clsx(
+        "bg-bg-subtle p-4 flex flex-col gap-4 rounded-md",
+        className,
+      )}
+    >
+      <div className="flex justify-between items-center">
+        <Heading level="3" size="xsmall">
+          {tittel}
+        </Heading>
+        <Detail className="uppercase flex items-center">{kilde}</Detail>
+      </div>
+      {children}
+    </div>
+  );
+};
